requirements.txt
Django==2.2
django-boogie~=0.11.1
psycopg2
Pillow
pycpfcnpj
